requirements.txt
matplotlib
numpy
scipy
Cython
tqdm
bridson
easydict
scikit-image
Pillow
imgaug==0.2.6
tensorboard
graphviz
opencv-python-headless
albumentations
pytorch
torchvision==0.4.1